Modify error reporting on connecting to API
If the initial connection fails, this can have multiple errors:
Either the endpoint is not correct, in which case we take another
endpoint anyway and try again.
Or we have a real error using any of the specified endpoints, in which
case we want to know WHY the connection could not be established.
Therefore, we store the last raised connection error and if we still
have no valid proxy at the end, we can log why everything failed.