Refactored Iterators because of empty Filter bug
In all iterators, there was an error if the filter loop was not entered,
because the user did not specify some filter params. In that case,
no results would be yield from the iterators, resulting in None values
in the getter() methods.

Also improved iterations over TestSuites with the recursive option.
Now, even if the current TestSuite does not match the search parameters,
all sub suites (and only these) are iterated for a possible match.
Before that change, if a parent TestSuite does not match, so does
any child suite.