Fix error on TestPlan.iterTestCase()
We have to get both types of TestCase, one from the TestPlan and the TestCase
itself because the missing step information. Therefore, we have too many
arguments which describe the ID and the Version ID of the TestCase.
So the IDs of one type has to be deleted before merging.