Workaround for API Error on getTestCasesForTestPlan
The 'exeucution_status' argument ist not interpreted correctly in Testlink 1.9.11.
Other versions were not checked, but to be sure, we ignore this argument
for the raw API call.

We still can filter for the execution status afterwards. This takes longer,
but it's better than not being able to filter it at all.