Fix error on ThreadSafeTransport of xmlrpclib ServerProxy
In Python <= 2.6 a HTTP connection has been established each time
a call was made to the server. In Python >= 2.7 however, the
Transport class tries to cache connections per host to reduce
overhead and to profit from the HTTP Keep-Alive of the Server.

But when a connection is reused every time, we get into trouble
if we use multiple threads for API calls. So we have to disable
the internal cache of the current Transport instance by simply
resetting it to None each time we try to make a connection.