Modify reconnection behaviour
Now if we get any(!) exception or socket.error at _query(), we try
to reconnect to Testlink.

If this reconnect (or initial connect) fails, we wait some seconds
before we try to reconnect and resend the request.

With this construct, we can loose connection right within a runnig
query and recover from it within a time span of 25 seconds
(5 reconnection attempts with 5 seconds waiting time).