Move actual build() method to builder instead of api fromAPIBuilder
This is because if we use the builder to build objects, which have other
objects as parents, we need slots in the builder to store these parents.

The fromAPIBuilder does not have such slots, which makes it hard to
build the object there.

Also if we want to sanity check for parent objects, this can also not
be done in the fromAPIBuilder.